Reflect subchapters in outline
This is only possible to a certain extent, because some chapter PDFs already have an outline. I chose to leave that outline intact if present, but that means the respective chapters can only be children of root. It is not easily possible to copy the outline at full depth or to append a chapter PDF with a parent other than root/top-level.